Pinhole camera

+++ simple/src/random.rs
@@ -6,9 +6,9 @@ pub fn random_in_unit_sphere() -> Vec3 {
     loop {
         let p = 2.0
             * vec3(
-                rng.gen_range(0.0 as f32, 1.0),
-                rng.gen_range(0.0 as f32, 1.0),
-                rng.gen_range(0.0 as f32, 1.0),
+                rng.gen_range(0.0_f32, 1.0),
+                rng.gen_range(0.0_f32, 1.0),
+                rng.gen_range(0.0_f32, 1.0),
             );
         if dot(p, p) < 1.0 {
             break p;
@@ -18,5 +18,5 @@ pub fn random_in_unit_sphere() -> Vec3 {
 
 pub fn random_float() -> f32 {
     let mut rng = rand::thread_rng();
-    rng.gen_range(0.0 as f32, 1.0)
+    rng.gen_range(0.0_f32, 1.0)
 }
